Split tests into many files
We can achieve better concurrency with smaller files, as they can be balanced over more processes.
