+++ test/spec/php_5.6_newrelic_spec.rb
@@ -1,4 +1,4 @@
-require_relative "php_shared"
+require_relative "php_shared_newrelic"
 
 describe "A PHP 5.6 application with a composer.json", :requires_php_on_stack => "5.6" do
 	include_examples "A PHP application with a composer.json", "5.6"
