+++ test/spec/php_7.2_newrelic_spec.rb
@@ -1,4 +1,4 @@
-require_relative "php_shared"
+require_relative "php_shared_newrelic"
 
 describe "A PHP 7.2 application with a composer.json", :requires_php_on_stack => "7.2" do
 	include_examples "A PHP application with a composer.json", "7.2"
